02-samples/11-customer-support-solar-panel/requirements.txt
boto3
botocore
awscli
opensearch-py
requests-aws4auth
pyyaml
retrying
pandas
strands-agents
strands-agents-tools
mem0ai
python-dotenv
mem0ai
langfuse
atlassian-python-api